Add EQS in Switzerland
Added a few EQs in Switzerland from the catalog of the Swiss
Seismological Service to be able to generate an EQ hazard event set for
Switzerland (see readme for details).
Feel free to change the name of this file to a longer one
(e.g. "_extended_CHE.csv") and store it separately if you do not want to
use it as the standard CLIMADA EQ database file.